refactor: move RenameAdapter to new file
